app/cnn/requirements.txt
# Deep Learning
# Note: TensorFlow requires Python 3.9-3.12
# If you have Python 3.13+, you may need to use an older Python version
tensorflow>=2.14.0; python_version < '3.13'
tensorflow>=2.15.0; python_version >= '3.13'
keras>=2.14.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
